Split parsing for specs 1.4 and 1.5
