+++ lib/parse-collada/src/lib.rs
@@ -60,483 +60,16 @@
 
 extern crate xml;
 
+pub use v1_5::*;
 pub use xml::common::TextPosition;
 pub use xml::reader::Error as XmlError;
 
-use std::io::Read;
 use std::fmt::{self, Display, Formatter};
-use xml::attribute::OwnedAttribute;
 use xml::common::Position;
-use xml::name::OwnedName;
-use xml::namespace::Namespace;
-use xml::reader::{EventReader, ParserConfig};
-use xml::reader::XmlEvent::*;
 
-pub static COLLADA_ATTRIBS: &'static [&'static str] = &["version", "xmlns", "base"];
-static PARSER_CONFIG: ParserConfig = ParserConfig {
-    trim_whitespace: true,
-    whitespace_to_characters: true,
-    cdata_to_characters: true,
-    ignore_comments: true,
-    coalesce_characters: true,
-};
-
-/// Represents a parsed COLLADA document.
-#[derive(Debug, Clone)]
-pub struct Collada {
-    /// The version string for the COLLADA specification used by the document.
-    ///
-    /// Only "1.4.0", "1.4.1", and "1.5.0" are supported currently.
-    pub version: String,
-
-    /// Global metadata about the COLLADA document.
-    pub asset: Asset,
-
-    /// The base uri for any relative URIs in the document.
-    ///
-    /// Specified by the `base` attribute on the root `<COLLADA>` element.
-    pub base_uri: Option<AnyUri>,
-}
-
-impl Collada {
-    /// Read a COLLADA document from a string.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # #![allow(unused_variables)]
-    /// use parse_collada::Collada;
-    ///
-    /// static DOCUMENT: &'static str = r#"
-    ///     <?xml version="1.0" encoding="utf-8"?>
-    ///     <COLLADA xmlns="http://www.collada.org/2005/11/COLLADASchema" version="1.4.1">
-    ///         <asset />
-    ///     </COLLADA>
-    /// "#;
-    ///
-    /// let collada = Collada::from_str(DOCUMENT).unwrap();
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns `Err` if the document is invalid or malformed in some way. For details about
-    /// COLLADA versions, 3rd party extensions, and any other details that could influence how
-    /// a document is parsed see the [crate-level documentation][crate].
-    ///
-    /// [crate]: index.html
-    pub fn from_str(source: &str) -> Result<Collada> {
-        let reader = EventReader::new_with_config(source.as_bytes(), PARSER_CONFIG.clone());
-        Collada::parse(reader)
-    }
-
-    /// Attempts to parse the contents of a COLLADA document.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # #![allow(unused_variables)]
-    /// use std::fs::File;
-    /// use parse_collada::Collada;
-    ///
-    /// let file = File::open("resources/blender_cube.dae").unwrap();
-    /// let collada = Collada::read(file).unwrap();
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns `Err` if the document is invalid or malformed in some way. For details about
-    /// COLLADA versions, 3rd party extensions, and any other details that could influence how
-    /// a document is parsed see the [crate-level documentation][crate].
-    ///
-    /// [crate]: index.html
-    pub fn read<R: Read>(reader: R) -> Result<Collada> {
-        let reader = EventReader::new_with_config(reader, PARSER_CONFIG.clone());
-        Collada::parse(reader)
-    }
-
-    /// The logic behind parsing the COLLADA document.
-    ///
-    /// `from_str()` and `read()` just create the `xml::EventReader` and then defer to `parse()`.
-    fn parse<R: Read>(mut reader: EventReader<R>) -> Result<Collada> {
-        let mut collada = Collada {
-            version: String::new(),
-            asset: Asset::default(),
-            base_uri: None,
-        };
-
-        // Eat the `StartDocument` event. It has no useful information for our purposes, but it
-        // will always be the first event emitted, even if there's no XML declaration at the
-        // beginning of the document. This is defined as part of the xml-rs API as of v0.3.5,
-        // but it's possible this can will change in the future.
-        match reader.next()? {
-            StartDocument { .. } => {},
-            _ => panic!("First event from EventReader wasn't StartDocument"),
-        }
-
-        // The next element will always be the `<COLLADA>` tag. This will specify what version of
-        // the COLLADA spec is being used, which is how we'll determine our sub-parser.
-        let (name, attributes) = match reader.next()? {
-            StartElement { name, attributes, namespace: _ } => {
-                // If the element isn't the `<COLLADA>` tag then the document is malformed,
-                // return an error.
-                if name.local_name != "COLLADA" {
-                    return Err(Error {
-                        position: reader.position(),
-                        kind: ErrorKind::UnexpectedRootElement {
-                            element: name.local_name,
-                        }
-                    })
-                }
-
-                (name, attributes)
-            }
-
-            // I'm *almost* 100% certain that the only event that can follow the `StartDocument`
-            // event is a `StartElement` event. As of v0.3.5, xml-rs doesn't support
-            // `<!DOCTYPE>` or processing instructions, and it ignores whitespace and comments
-            // (according to how we configure the parser), and those are the only things allowed
-            // between `StartDocument` and the first `StartElement`. If xml-rs changes its
-            // behavior this will need to be updated.
-            event @ _ => { panic!("Unexpected event: {:?}", event); }
-        };
-
-        // Valide the attributes on the `<COLLADA>` tag.
-        // Use boolean flags to track if specific attributes have been encountered.
-        let mut has_found_version = false;
-        let mut has_found_base = false;
-
-        for attribute in attributes {
-            // NOTE: I'm using `if` blocks instead of `match` here because using `match`
-            // won't allow for the name to be moved out of `attribute`. Using `if` saves
-            // some unnecessary allocations. I expect at some point Rust will get smart
-            // enough that this will no longer be an issue, at which point we should
-            // change this to use `match`, as that keeps better with Rust best practices.
-            if attribute.name.local_name == "version" {
-                if !has_found_version {
-                    collada.version = attribute.value;
-                    has_found_version = true;
-                } else {
-                    return Err(Error {
-                        position: reader.position(),
-                        kind: ErrorKind::UnexpectedAttribute {
-                            element: name.local_name,
-                            attribute: "version".to_owned(),
-                            expected: COLLADA_ATTRIBS.into(),
-                        },
-                    })
-                }
-            } else if attribute.name.local_name == "base" {
-                if !has_found_base {
-                    // TODO: Do we need to validate the URI?
-                    collada.base_uri = Some(AnyUri(attribute.value));
-                    has_found_base = true;
-                } else {
-                    return Err(Error {
-                        position: reader.position(),
-                        kind: ErrorKind::UnexpectedAttribute {
-                            element: name.local_name,
-                            attribute: "base".to_owned(),
-                            expected: COLLADA_ATTRIBS.into(),
-                        },
-                    })
-                }
-            } else {
-                return Err(Error {
-                    position: reader.position(),
-                    kind: ErrorKind::UnexpectedAttribute {
-                        element: name.local_name,
-                        attribute: attribute.name.local_name,
-                        expected: COLLADA_ATTRIBS.into(),
-                    },
-                })
-            }
-        }
-
-        // Verify that all required attributes have been found.
-        if !has_found_version {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::MissingAttribute {
-                    element: name.local_name,
-                    attribute: "version",
-                },
-            })
-        }
-
-        // The next event must be the `<asset>` tag. No text data is allowed, and
-        // whitespace/comments aren't emitted.
-        let (_name, _, _) = required_start_element(&mut reader, "COLLADA", "asset")?;
-        collada.asset = parse_asset(&mut reader)?;
-
-        // Eat any events until we get to the `</COLLADA>` tag.
-        // TODO: Actually parse the body of the document.
-        loop {
-            match reader.next()? {
-                EndElement { ref name } if name.local_name == "COLLADA" => { break }
-                _ => {}
-            }
-        }
-
-        // TODO: Verify the next event is the `EndDocument` event.
-        match reader.next()? {
-            EndDocument => {}
-
-            // Same logic here as with the starting event. The only things that can come after the
-            // close tag are comments, white space, and processing instructions, all of which we
-            // ignore. This can change with future versions of xml-rs, though.
-            event @ _ => { panic!("Unexpected event: {:?}", event); }
-        }
-
-        Ok(collada)
-    }
-}
-
-fn required_start_element<R: Read>(
-    reader: &mut EventReader<R>,
-    parent: &str,
-    search_name: &'static str,
-) -> Result<(OwnedName, Vec<OwnedAttribute>, Namespace)> {
-    match reader.next()? {
-        StartElement { name, attributes, namespace } => {
-            if search_name != name.local_name {
-                return Err(Error {
-                    position: reader.position(),
-                    kind: ErrorKind::UnexpectedElement {
-                        element: name.local_name,
-                        parent: parent.into(),
-                        expected: vec![search_name],
-                    },
-                })
-            }
-
-            return Ok((name, attributes, namespace));
-        }
-
-        EndElement { name } => {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::MissingElement {
-                    expected: search_name,
-                    parent: name.local_name,
-                },
-            })
-        }
-
-        Characters(data) => {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::UnexpectedCharacterData {
-                    element: parent.into(),
-                    data: data,
-                }
-            })
-        }
-
-        ProcessingInstruction { .. } => { unimplemented!(); }
-
-        event @ _ => { panic!("Unexpected event: {:?}", event); }
-    }
-}
-
-fn optional_start_element<R: Read>(
-    reader: &mut EventReader<R>,
-    parent: &str,
-    search_names: &'static [&'static str],
-    current_element: usize,
-) -> Result<Option<(OwnedName, Vec<OwnedAttribute>, Namespace)>> {
-    let current_search_names = &search_names[current_element ..];
-
-    match reader.next()? {
-        StartElement { name, attributes, namespace } => {
-            if !current_search_names.contains(&&*name.local_name) {
-                return Err(Error {
-                    position: reader.position(),
-                    kind: ErrorKind::UnexpectedElement {
-                        element: name.local_name,
-                        parent: parent.into(),
-                        expected: search_names.into(),
-                    },
-                })
-            }
-
-            return Ok(Some((name, attributes, namespace)));
-        }
-
-        EndElement { .. } => {
-            return Ok(None);
-        }
-
-        Characters(data) => {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::UnexpectedCharacterData {
-                    element: parent.into(),
-                    data: data,
-                }
-            })
-        }
-
-        ProcessingInstruction { .. } => { unimplemented!(); }
-
-        event @ _ => { panic!("Unexpected event: {:?}", event); }
-    }
-}
-
-fn text_only_element<R: Read>(
-    reader: &mut EventReader<R>,
-    parent: &str,
-) -> Result<Option<String>> {
-    match reader.next()? {
-        Characters(data) => {
-            end_element(reader, parent)?;
-            return Ok(Some(data))
-        }
-
-        StartElement { name, attributes: _, namespace: _ } => {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::UnexpectedElement {
-                    element: name.local_name,
-                    parent: parent.into(),
-                    expected: vec![],
-                },
-            })
-        }
-
-        EndElement { .. } => {
-            return Ok(None);
-        }
-
-        ProcessingInstruction { .. } => { unimplemented!(); }
-
-        event @ _ => { panic!("Unexpected event: {:?}", event); }
-    }
-}
-
-fn end_element<R: Read>(reader: &mut EventReader<R>, parent: &str) -> Result<()> {
-    match reader.next()? {
-        EndElement { .. } => {
-            return Ok(());
-        }
-
-        StartElement { name, attributes: _, namespace: _ } => {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::UnexpectedElement {
-                    element: name.local_name,
-                    parent: parent.into(),
-                    expected: vec![],
-                },
-            })
-        }
-
-        Characters(data) => {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::UnexpectedCharacterData {
-                    element: parent.into(),
-                    data: data,
-                }
-            })
-        }
-
-        ProcessingInstruction { .. } => { unimplemented!(); }
-
-        event @ _ => { panic!("Unexpected event: {:?}", event); }
-    }
-}
-
-fn parse_asset<R: Read>(reader: &mut EventReader<R>) -> Result<Asset> {
-    let mut asset = Asset::default();
-
-    // Parse the children of the `<asset>` tag.
-    static ASSET_CHILDREN: &'static [&'static str] = &["contributor"];
-    while let Some((_name, attributes, _)) = optional_start_element(reader, "asset", ASSET_CHILDREN, 0)? {
-        let contributor = parse_contributor(reader, attributes)?;
-        asset.contributors.push(contributor);
-    }
-
-    Ok(asset)
-}
-
-fn parse_contributor<R: Read>(reader: &mut EventReader<R>, attributes: Vec<OwnedAttribute>) -> Result<Contributor> {
-    // Make sure the `<contributor>` element has no attributes.
-    if attributes.len() != 0 {
-        return Err(Error {
-            position: reader.position(),
-            kind: ErrorKind::UnexpectedAttribute {
-                element: "contributor".into(),
-                attribute: attributes[0].name.local_name.clone(),
-                expected: vec![],
-            },
-        })
-    }
-
-    let mut contributor = Contributor::default();
-
-    static EXPECTED_ELEMENTS: &'static [&'static str] = &[
-        "author",
-        "authoring_tool",
-        "comments",
-        "copyright",
-        "source_data",
-    ];
-
-    fn verify_attributes<R: Read>(reader: &EventReader<R>, name: &OwnedName, attributes: Vec<OwnedAttribute>) -> Result<()> {
-        // Make sure the child element has no attributes.
-        if attributes.len() != 0 {
-            return Err(Error {
-                position: reader.position(),
-                kind: ErrorKind::UnexpectedAttribute {
-                    element: name.local_name.clone(),
-                    attribute: attributes[0].name.local_name.clone(),
-                    expected: vec![],
-                },
-            })
-        }
-
-        Ok(())
-    }
-
-    let mut current_element = 0;
-    while let Some((element_name, element_attributes, _)) = optional_start_element(reader, "contributor", EXPECTED_ELEMENTS, current_element)? {
-        match &*element_name.local_name {
-            "author" => {
-                verify_attributes(reader, &element_name, element_attributes)?;
-                contributor.author = text_only_element(reader, "author")?
-            }
-
-            "authoring_tool" => {
-                verify_attributes(reader, &element_name, element_attributes)?;
-                contributor.authoring_tool = text_only_element(reader, "authoring_tool")?;
-            }
-
-            "comments" => {
-                verify_attributes(reader, &element_name, element_attributes)?;
-                contributor.comments = text_only_element(reader, "authoring_tool")?;
-            }
-
-            "copyright" => {
-                verify_attributes(reader, &element_name, element_attributes)?;
-                contributor.copyright = text_only_element(reader, "copyright")?;
-            }
-
-            "source_data" => {
-                verify_attributes(reader, &element_name, element_attributes)?;
-                contributor.source_data = text_only_element(reader, "source_data")?.map(Into::into);
-            }
-
-            _ => { panic!("Unexpected element name: {}", element_name); }
-        }
-
-        current_element = EXPECTED_ELEMENTS
-            .iter()
-            .position(|&name| name == element_name.local_name)
-            .expect("Element wasn't in expected elements");
-    }
-
-    Ok(contributor)
-}
+mod utils;
+mod v1_4;
+mod v1_5;
 
 /// A COLLADA parsing error.
 ///
@@ -661,6 +194,10 @@ pub enum ErrorKind {
         data: String,
     },
 
+    UnsupportedVersion {
+        version: String,
+    },
+
     /// The XML in the document was malformed in some way.
     XmlError(XmlError),
 }
@@ -702,6 +239,10 @@ impl Display for ErrorKind {
 
             ErrorKind::UnexpectedCharacterData { ref element, data: _ } => {
                 write!(formatter, "<{}> contained non-markup text data which isn't allowed", element)
+            }
+
+            ErrorKind::UnsupportedVersion { ref version } => {
+                write!(formatter, "Unsupported COLLADA version {:?}, supported versions are \"1.4.0\", \"1.4.1\", \"1.5.0\"", version)
             }
 
             ErrorKind::XmlError(ref error) => {
@@ -755,57 +296,4 @@ impl<'a> Display for StringListDisplay<'a> {
 
         Ok(())
     }
-}
-
-/// Asset-management information about an element.
-///
-/// Includes both asset metadata, such as a list of contributors and keywords, as well
-/// as functional information, such as units of distance and the up axis for the asset.
-///
-/// # COLLADA Versions
-///
-/// `coverage` and `extras` were added in COLLADA version `1.5.0`.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct Asset {
-    /// The list of contributors who worked on the asset.
-    pub contributors: Vec<Contributor>,
-}
-
-/// Information about a contributor to an asset.
-///
-/// Contributor data is largely free-form text data meant to informally describe either the author
-/// or the author's work on the asset. The exceptions are `author_email`, `author_website`, and
-/// `source_data`, which are strictly formatted data (be it a URI or email address).
-///
-/// # COLLADA Versions
-///
-/// `author_email` and `author_website` were added in COLLADA version `1.5.0`.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct Contributor {
-    /// The author's name, if present.
-    pub author: Option<String>,
-
-    /// The author's full email address, if present.
-    // TODO: Should we use some `Email` type? The 1.5.0 COLLADA spec provides an RFC defining the
-    // exact format this data follows (I assume it's just the RFC that defines valid email
-    // addresses).
-    pub author_email: Option<String>,
-
-    /// The URL for the author's website, if present.
-    pub author_website: Option<AnyUri>,
-
-    /// The name of the authoring tool.
-    pub authoring_tool: Option<String>,
-
-    /// Free-form comments from the author.
-    pub comments: Option<String>,
-
-    /// Copyright information about the asset. Does not adhere to a formatting standard.
-    pub copyright: Option<String>,
-
-    /// A URI reference to the source data for the asset.
-    ///
-    /// For example, if the asset based off a file `tank.s3d`, the value might be
-    /// `c:/models/tank.s3d`.
-    pub source_data: Option<AnyUri>,
 }
